requirements/req-dev.txt
asgiref==3.2.3
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.3
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.5.4
django-cities==0.6
django-ckeditor==5.7.1
django-countries==6.1.2
django-filter==2.2.0
django-js-asset==1.2.2
django-map-widgets==0.3.0
django-mathfilters==0.4.0
django-model-utils==4.0.0
django-pdfkit==0.3.1
django-rest-passwordreset==1.1.0
django-user-agents==0.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.9
Markdown==3.1.1
mysqlclient==1.4.6
netifaces==0.10.6
numpy==1.18.2
pandas==1.0.1
pdfkit==0.6.0
Pillow==7.0.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
swapper==1.1.2.post1
tqdm==4.46.1
ua-parser==0.10.0
urllib3==1.25.9
user-agents==2.1
xlrd==1.2.0
django-modeladmin-reorder
oauth2client
firebase-admin